fyp-backend-main/fyp-backend-main/execution-service/requirements.txt
flask
flask-cors
pymongo
requests
python-dotenv
boto3
pyyaml